New incremental fetching hook
I was running into issues with Heroku timing out HTTP requests after
30 seconds, so instead of one large post request with all of the user
songs, this only sends 20 at a time to finish before 30 seconds. Once
all lyrics have been fetched, they are posted to "/update-database"
where they are saved to MongoDB. It also aggregates all of the updated
lyrics so that the global lyrics data is up to date as well. The user
lyrics, global lyrics, and userID context variables are set from the
response of this request.